// src/models/sicap/Menu.ts
/** Tabla dbo.menus (legacy SICAP) */
export interface Menu {
  corr_men: number;
  cod_men: string | null;
  Nom_men: string | null;
  Tip_men: string | null;
  acc_men: string | null;
  tar_men: string | null;
  imagen: string | null;
  target_net: string | null;
  id_tip: number | null;
  url_net: string | null;
  url_spe: string | null;
  asp: string | null;
}